requirements.txt
torch>=2.0.0
torchinfo
torchvision>=0.15.0
numpy>=1.23.0
matplotlib>=3.7.0
Pillow>=9.5.0
tqdm>=4.65.0
